Fixed Tools
Fixed how the hotkeys were being loaded when an item on the combo box
was selected.

Added functionality for tools to be loaded dynamically like environments
already were.

Added Event hookup for tools being clicked to actually load the tools.

Added event to reload the context menu when the Edit screen is closed so
accurate/up-to-date data is displayed.