Properly Dispose OnClosing
Only the hotkey monitor was disposing and registry monitor was stopping
when the Exit button was clicked.
Now I added OnClosing event to catch other types of closing events and
added disposing of the registry monitor, disposing of the
ContextMenuStrip and disposing of the Icon.

The biggest noticable change is that NOW when you close the program, the
Icon will be removed from the systray without having to hover over it.